Version 1.3.4
09/06/2020, Bing Li

An important update is done.

The ClusterServerContainer/ClusterChildContainer has the capability to leaves one joined cluster and joins another one.

This is useful when one cluster lacks in computing resources or its computing resources are superfluous.

The current lines of code are 203635.

The number of files is 3040.

================================================

09/04/2020, Bing Li

Now I need to implement a cluster that is able to scale automatically according to its current workload.

That is an important feature.

The potential solution is discussed as follows.

(The solution is written on the way to Baoji to mourn President Li. 09/05/2020, LB)

1) Two clusters are established.

2) The first one works on a task. The cluster is called the task cluster.

3) The second one plays the role of the pool of additional computing nodes. The cluster is called the pool cluster.

4) When the task cluster cannot afford the workload, it sends a request to the pool cluster.

5) Then, the pool cluster selects one child from all of the children in the cluster.

6) The selected child is excluded from the children list of the root of the pool cluster.

7) The selected child is sent a join notification to the root of the task cluster.

8) Thereafter, the scale of the task cluster is increased to take more workload.