Version 1.3.7.0
09/12/2020, Bing Li

The version is updated for the scalability.

Usually, the task cluster is a load-balancing system.

So, if the workload is heavy, the entire children should become idle and the same number of children selected from the pool cluster join the task cluster to accomplish the new tasks.

Now the feature is implemented.

One important update is the broadcasting within a fixed number of children.

But upper level issues should be taken care by programmers, especially for the state management.

The current lines of code are 205165.

The number of files is 3068.