Version 1.3.9.0
08/28/2021, Bing Li

ServerProfile is not initialized properly.

When distributed nodes are deployed on different computing devices, the registry server IP needs to be retained in ServerProfile.

But since I did not test the case, the registry server IP is always 127.0.0.1.

The bug is fixed.

The current lines of code are 205916.

The number of files is 3084.

================================================

08/25/2021, Bing Li

I attempt to change the SP pattern to replace the while-true loop with the notify/wait structure in ClusterChildContainer and ClusterServerContainer.

	TerminateSignal.SIGNAL().notifyAllTermination();

With the updating of notify/wait, the line is not needed. But to keep compatible, just leave the below line for some time.

	TerminateSignal.SIGNAL().setTerminated();

The current lines of code are 205895.

The number of files is 3084.