Version 1.4.3.0
04/22/2022, Bing Li

After testing initially, it is unreasonable to send the notification, InitReadFeedbackNotification, back to the client.

I will test the issue further.

The current lines of code are 211398.

The number of files is 3152.

================================================

04/22/2022, Bing Li

One big bug is fixed.

The InitReadFeedbackThread always blocks and consumes 25% CPU usage after the interaction starts between the server and the client.

VisualVM is used to detect the problem.

I need to use the tool frequently.

I need to ensure whether the thread is useful or not.

The thread is constructed when I am novice of distributed programming.

================================================

04/13/2022, Bing Li

The system messages are defined in the below classes.

org.greatfree.message.SystemMessageType

org.greatfree.message.multicast.MulticastMessageType

org.greatfree.concurrency.threading.message.ATMMessageType

The largest number is 119.

================================================

04/09/2022, Bing Li

The class, org.greatfree.concurrency.Async, is updated.

One additional method is added to adapt to the operations upon cryptography algorithms.

The current lines of code are 210967.

The number of files is 3151.

================================================

04/06/2022, Bing Li

I need to implement the cryptography cluster.

Some fundamental APIs as follows need to be updated.

	org.greatfree.server.container.Peer

	org.greatfree.client.CSClient

The current lines of code are 210794.

The number of files is 3151.

================================================

02/09/2022, Bing Li

I am trying to update the Distributed Primitives with the NIO networking to raise the quality.

This is an important change to the Wind because the DP is the foundation of the Wind.

Now the notification-based primitives are done initially.

The request-based ones should be updated in the same way.

The current lines of code are 210781.

The number of files is 3151.