Version 1.4.4.7
05/03/2022, Bing Li

The multicasting with the low-level message, ServerMessage, is finally done for WindCry.

One trivial bug takes me one night, i.e., disposing the peer.

In the multicasting structure, multiple peers participate.

Some of them leave earlier such that exceptions might be caused when later ones are disposed.

It should not affect later peers' disposing their own resources, such as the dispatchers.

Now the bugs are fixed.

The current lines of code are 211589.

The number of files is 3155.